Add Navigation Controller:
Couldn’t make the push to the tweet view to work from the containerView of menuVC.

Ended up making the timeline as rootVC in the appDelegate and then I’m able to push.

What I would do is add a tabBar below that lets us select Profile, TimeLine and Mentions buttons that are missing now.